Tune gcc -Os to produce smaller code for AVR32.
These patches tune the optimization system to produce about 3% smaller code
when -Os is specified. In gcc >=4.3 this is considered an optimization bug
(see  http://gcc.gnu.org/bugzilla/show_bug.cgi?id=30908) whereas in gcc-4.2
it is just a question of tuning the inliner to the size of the AVR32
function call overhead.